+++ DSView/pv/dialogs/deviceoptions.cpp
@@ -705,7 +705,7 @@ void DeviceOptions::analog_probes(QGridLayout &layout)
         probe_widget->setLayout(probe_layout); 
 
         bool ch_enabled = probe->enabled;
-        if (ch_dex < _lst_probe_enabled_status.size()){
+        if ((size_t)ch_dex < _lst_probe_enabled_status.size()){
             ch_enabled = _lst_probe_enabled_status[ch_dex];
         }
